requirements.txt
aiosqlite
python-dotenv
httpx
openai
pypdf
reportlab
